docs(wheel-picker): upgrade example to use form
