+++ apps/v4/src/content/docs/components/wheel-picker.mdx
@@ -60,6 +60,6 @@ import { WheelPicker, WheelPickerWrapper } from "@/components/ui/wheel-picker";
 
 <ComponentPreview name="wheel-picker-demo" />
 
-### Time Picker
+### Form
 
-<ComponentPreview name="wheel-picker-time-picker" />
+<ComponentPreview name="wheel-picker-form" />
